refactor: enum패키지명을 constant로 변경
- 해당 패키지에 enum과 constant class가 함꼐 있어서 이름을 변경험
