+++ src/main/java/lotto/constant/LottoConstant.java
@@ -1,11 +1,10 @@
-package lotto.enums;
+package lotto.constant;
 
 public final class LottoConstant {
     public static final int MIN_LOTTO_NUMBER = 1;
     public static final int MAX_LOTTO_NUMBER = 45;
     public static final int LOTTO_NUMBER_COUNT = 6;
     public static final int PRICE = 1000;
-    public static final int LOTTO_RANK_COUNT = 5;
 
     private LottoConstant() {
     }
