+++ src/main/java/lotto/model/result/LottoResult.java
@@ -2,7 +2,7 @@ package lotto.model.result;
 
 import java.util.EnumMap;
 import java.util.List;
-import lotto.enums.LottoWinInfo;
+import lotto.constant.LottoWinInfo;
 import lotto.model.draw.DrawNumber;
 import lotto.model.draw.LottoNumbers;
 import lotto.model.purchase.Lotto;
