+++ src/main/java/lotto/view/LottoResultView.java
@@ -2,7 +2,7 @@ package lotto.view;
 
 import java.util.EnumMap;
 import java.util.Map.Entry;
-import lotto.enums.LottoWinInfo;
+import lotto.constant.LottoWinInfo;
 import lotto.model.result.Profit;
 
 public class LottoResultView {
